Move the initialization to application.
