+++ testapps/testapp/src/main/java/com/microsoft/identity/client/testapp/MsalSampleApp.java
@@ -24,9 +24,17 @@
 package com.microsoft.identity.client.testapp;
 
 import android.app.Application;
+import android.util.Log;
 
 import com.microsoft.identity.client.ILoggerCallback;
 import com.microsoft.identity.client.Logger;
+import com.microsoft.identity.common.internal.telemetry.Telemetry;
+import com.microsoft.identity.common.internal.telemetry.observers.ITelemetryAggregatedObserver;
+import com.microsoft.identity.common.internal.telemetry.observers.ITelemetryDefaultObserver;
+
+import java.util.Iterator;
+import java.util.List;
+import java.util.Map;
 
 /**
  * MSAL sample app.
@@ -56,6 +64,34 @@ public class MsalSampleApp extends Application {
                 mLogs.append(message).append('\n');
             }
         });
+
+        Telemetry.getInstance().addObserver(new TelemetryAggregatedObserver());
+        Telemetry.getInstance().addObserver(new TelemetryDefaultObserver());
+        Telemetry.getInstance().removeObserver(TelemetryDefaultObserver.class);
+    }
+
+    class TelemetryDefaultObserver implements ITelemetryDefaultObserver {
+        public void onReceived(List<Map<String, String>> telemetryData) {
+            for (Map properties : telemetryData) {
+                final Iterator iterator = properties.entrySet().iterator();
+                while (iterator.hasNext()) {
+                    final Map.Entry pair = (Map.Entry) iterator.next();
+                    Log.e("Test", pair.getKey() + ":" + pair.getValue());
+                }
+                Log.e("Test", "====================================" + '\n');
+            }
+        }
+    }
+
+    class TelemetryAggregatedObserver implements ITelemetryAggregatedObserver {
+        public void onReceived(Map<String, String> telemetryData) {
+            final Iterator iterator = telemetryData.entrySet().iterator();
+            while (iterator.hasNext()) {
+                final Map.Entry pair = (Map.Entry) iterator.next();
+                Log.e("Test", pair.getKey() + ":" + pair.getValue());
+            }
+            Log.e("Test", "====================================" + '\n');
+        }
     }
 
     String getLogs() {
